feat(logger): introduce logger utility with configurable log levels refactor(helper imports): update import paths to use new logger module docs: update documentation for logger usage and configuration
